refactoring: models,  implemented classes in controllers
