+++ src/controllers/handlers/auth_handler.py
@@ -1,15 +1,16 @@
 '''Handlers related to Authentication'''
 
-import hashlib
 import logging
 from typing import List, Tuple
 
-from config.message_prompts import DisplayMessage, Prompts, LogMessage
+from config.message_prompts import DisplayMessage, Headers, LogMessage, Prompts
 from config.queries import Queries
-from controllers import auth_controller as Authenticate
+from config.regex_patterns import RegexPattern
+from controllers.auth_controller import AuthController
 from database.database_access import DatabaseAccess as DAO
 from utils import validations
 from utils.custom_error import LoginError
+from utils.password_hasher import hash_password
 
 logger = logging.getLogger(__name__)
 
@@ -19,11 +20,21 @@ def handle_login() -> List[Tuple]:
 
     logger.debug(LogMessage.LOGIN_INITIATED)
     print(DisplayMessage.LOGIN_MSG)
-
     attempts_remaining = Prompts.ATTEMPT_LIMIT
 
-    data = Authenticate.login()
-    while not data:
+    while True:
+        # validating credentials even during login to prevent any Injections
+        username = validations.regex_validator(
+            prompt='Enter your username: ',
+            regex_pattern=RegexPattern.USERNAME_PATTERN,
+            error_msg=DisplayMessage.INVALID_TEXT.format(Headers.USERNAME)
+        )
+        password = validations.validate_password(prompt='Enter your password: ')
+
+        data = AuthController().login(username, password)
+        if data:
+            break
+
         attempts_remaining -= 1
         print(DisplayMessage.REMAINING_ATTEMPTS_MSG.format(count=attempts_remaining))
 
@@ -31,8 +42,6 @@ def handle_login() -> List[Tuple]:
             print(DisplayMessage.LOGIN_ATTEMPTS_EXHAUST_MSG)
             logger.debug(LogMessage.LOGIN_ATTEMPTS_EXHAUSTED)
             return None
-
-        data = Authenticate.login()
 
     logger.debug(LogMessage.LOGIN_SUCCESS)
     return data
@@ -44,8 +53,35 @@ def handle_signup() -> str:
     logger.debug(LogMessage.SIGNUP_INITIATED)
     print(DisplayMessage.SIGNUP_MSG)
 
+    player_data = {}
+    player_data['name'] = validations.regex_validator(
+        prompt='Enter your name: ',
+        regex_pattern=RegexPattern.NAME_PATTERN,
+        error_msg=DisplayMessage.INVALID_TEXT.format(Headers.NAME)
+    ).title()
+    player_data['email'] = validations.regex_validator(
+        prompt='Enter your email: ',
+        regex_pattern=RegexPattern.EMAIL_PATTERN,
+        error_msg=DisplayMessage.INVALID_TEXT.format(Headers.EMAIL)
+    )
+    player_data['username'] = validations.regex_validator(
+        prompt='Create your username: ',
+        regex_pattern=RegexPattern.USERNAME_PATTERN,
+        error_msg=DisplayMessage.INVALID_TEXT.format(Headers.USERNAME)
+    )
+    password = validations.validate_password(prompt='Create your password: ')
+    confirm_password = ''
+
+    while True:
+        confirm_password =  validations.validate_password(prompt='Confirm Password: ')
+        if password != confirm_password:
+            print(DisplayMessage.CONFIRM_PSWD_FAIL_MSG)
+        else:
+            break
+
+    player_data['password'] = confirm_password
     try:
-        username = Authenticate.signup()
+        username = AuthController().signup(player_data)
     except LoginError as e:
         print(e)
         logger.warning(e)
@@ -73,7 +109,7 @@ def handle_first_login(username: str, is_password_changed: int) -> None:
             else:
                 break
 
-        hashed_password = hashlib.sha256(confirm_password.encode('utf-8')).hexdigest()
+        hashed_password = hash_password(confirm_password)
         is_password_changed = 1
 
         DAO.write_to_database(
