+++ src/models/user.py
@@ -1,11 +1,11 @@
 '''Contains classes for Super Admin, Admin, Player'''
 
-from abc import ABC, abstractmethod
+from abc import ABC
 from datetime import datetime, timezone
 from typing import Dict
 
-from config.queries import Queries
-from database.database_access import DatabaseAccess as DAO
+from models.database_saver import DatabaseSaver
+from models.user_manager import UserManager
 from utils import validations
 
 
@@ -21,49 +21,6 @@ class User(ABC):
         self.role = role
         self.is_password_changed = 0 if role == 'admin' else 1
         self.registration_date = datetime.now(timezone.utc).strftime('%Y-%m-%d')
-
-
-class DatabaseSaver(ABC):
-    '''Abstract interface for saving to the database'''
-
-    @abstractmethod
-    def save_to_database(self) -> None:
-        '''Abstract method to save to the database'''
-        pass
-
-
-class UserManager:
-    '''
-    Manager class responsible for saving users to the database.
-
-    Attributes:
-        user (User): The user object to be managed.
-    '''
-
-    def __init__(self, user: User) -> None:
-        self.user = user
-
-    def save_to_database(self) -> None:
-        '''
-        Saves the user and their credentials to the database.
-        '''
-        user_data = (
-            self.user.user_id,
-            self.user.name,
-            self.user.email,
-            self.user.role,
-            self.user.registration_date
-        )
-
-        credentials = (
-            self.user.user_id,
-            self.user.username,
-            self.user.password,
-            self.user.is_password_changed
-        )
-
-        DAO.write_to_database(Queries.INSERT_USER_DATA, user_data)
-        DAO.write_to_database(Queries.INSERT_CREDENTIALS, credentials)
 
 
 class SuperAdmin(User, DatabaseSaver):
