+++ src/utils/initialize_app.py
@@ -1,6 +1,5 @@
 '''Runs at the start to create super admin and other tables in database'''
 
-import hashlib
 import logging
 import os
 import sqlite3
@@ -13,6 +12,7 @@ from config.queries import InitializationQueries
 from database.database_access import DatabaseAccess as DAO
 from models.user import SuperAdmin
 from utils.custom_error import DuplicateEntryError
+from utils.password_hasher import hash_password
 
 dotenv_path = Path('.env')
 load_dotenv(dotenv_path=dotenv_path)
@@ -33,12 +33,10 @@ class Initializer:
         super_admin_data['name'] = os.getenv('SUPER_ADMIN_NAME')
         super_admin_data['email'] = os.getenv('SUPER_ADMIN_EMAIL')
         super_admin_data['username'] = os.getenv('SUPER_ADMIN_USERNAME')
-
         password = os.getenv('SUPER_ADMIN_PASSWORD')
-        hashed_password = hashlib.sha256(password.encode('utf-8')).hexdigest()
-
-        super_admin_data['password'] = hashed_password
+        super_admin_data['password'] = hash_password(password)
         super_admin = SuperAdmin(super_admin_data)
+
         try:
             super_admin.save_to_database()
         except sqlite3.IntegrityError as e:
